+++ packages/jsonresume-theme-cjean/src/components/Skills.tsx
@@ -11,21 +11,25 @@ export default function Skills({ skills, ...props }: SkillsProps) {
   if (!skills || skills.length === 0) return null;
 
   return (
-    <Section name={t("skills")} {...props}>
-      {skills.map((skill) => (
-        <div className="mb-4 break-inside-avoid" key={skill.name}>
-          <h3 className="text-xl font-bold tracking-tight text-gray-900 dark:text-white">
-            {skill.name}
-          </h3>
-          <ul className="my-3 flex flex-wrap gap-2 leading-8">
-            {skill.keywords?.map((keyword) => (
-              <li className="badge" key={keyword}>
-                {keyword}
-              </li>
-            ))}
-          </ul>
-        </div>
-      ))}
+    <Section sectionId="skills" name={t("skills")} {...props}>
+      <dl>
+        {skills.map((skill) => (
+          <div className="mb-4 break-inside-avoid" key={skill.name}>
+            <dt className="text-xl font-bold tracking-tight text-gray-900 dark:text-white">
+              {skill.name}
+            </dt>
+            <dd>
+              <ul className="my-3 flex flex-wrap gap-2 leading-8">
+                {skill.keywords?.map((keyword) => (
+                  <li className="badge" key={keyword}>
+                    {keyword}
+                  </li>
+                ))}
+              </ul>
+            </dd>
+          </div>
+        ))}
+      </dl>
     </Section>
   );
 }